requirements.txt
streamlit
twilio
openai-whisper
gTTS
googletrans==4.0.0-rc1
pydub
numpy
pandas
